// projects/typeschema-angular-editor/src/lib/model/Document.ts
/**
 * Document generated on 2021-03-16
 * {@link https://github.com/apioo}
 */

export interface Document {
    id?: number;
    sourceId?: number;
    sourceName?: string;
    sourceUserName?: string;
    userId?: number;
    userName?: string;
    name?: string;
    description?: string;
    stars?: number;
    spec?: any;
    insertDate?: string;
}